Fix display bug in CPU graph
When you clicked on another detector and then came back to cpu, the
graph would be displayed improperly because of a bug in how the data is
ordered. The view helpers need the data to be ordered in reverse order
so that they can be pushed onto an array, with the last point pushed
being the first point on the graph. by calling reverse() on the same
data I already graphed, I was causing it to be ordered incorrectly.